refactor: 자립영역 선택 로직 user -> mission 으로 변경
