+++ src/main/java/com/dash/leap/domain/user/controller/UserController.java
@@ -3,11 +3,9 @@ package com.dash.leap.domain.user.controller;
 import com.dash.leap.domain.user.controller.docs.UserControllerDocs;
 import com.dash.leap.domain.user.dto.request.ChatbotSettingRequest;
 import com.dash.leap.domain.user.dto.request.LoginRequest;
-import com.dash.leap.domain.user.dto.request.MissionAreaSettingRequest;
 import com.dash.leap.domain.user.dto.request.UserRegisterRequest;
 import com.dash.leap.domain.user.dto.response.ChatbotSettingResponse;
 import com.dash.leap.domain.user.dto.response.LoginResponse;
-import com.dash.leap.domain.user.dto.response.MissionAreaSettingResponse;
 import com.dash.leap.domain.user.dto.response.UserRegisterResponse;
 import com.dash.leap.domain.user.service.UserService;
 import com.dash.leap.global.auth.user.CustomUserDetails;
@@ -51,15 +49,6 @@ public class UserController implements UserControllerDocs {
     ) {
         ChatbotSettingResponse chatbotSettingResponse = userService.leapySetting(userDetails.user(), request);
         return ResponseEntity.ok(chatbotSettingResponse);
-    }
-
-    @PatchMapping("/mission-area")
-    public ResponseEntity<MissionAreaSettingResponse> missionAreaSetting(
-            @Valid @RequestBody MissionAreaSettingRequest request,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        MissionAreaSettingResponse settingResponse = userService.missionSetting(userDetails.user(), request);
-        return ResponseEntity.ok(settingResponse);
     }
 
     @PostMapping("/logout")
