+++ src/main/java/com/dash/leap/domain/user/controller/docs/UserControllerDocs.java
@@ -2,11 +2,9 @@ package com.dash.leap.domain.user.controller.docs;
 
 import com.dash.leap.domain.user.dto.request.ChatbotSettingRequest;
 import com.dash.leap.domain.user.dto.request.LoginRequest;
-import com.dash.leap.domain.user.dto.request.MissionAreaSettingRequest;
 import com.dash.leap.domain.user.dto.request.UserRegisterRequest;
 import com.dash.leap.domain.user.dto.response.ChatbotSettingResponse;
 import com.dash.leap.domain.user.dto.response.LoginResponse;
-import com.dash.leap.domain.user.dto.response.MissionAreaSettingResponse;
 import com.dash.leap.domain.user.dto.response.UserRegisterResponse;
 import com.dash.leap.global.auth.user.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
@@ -35,13 +33,6 @@ public interface UserControllerDocs {
     @ApiResponse(description = "설정 성공", responseCode = "200")
     ResponseEntity<ChatbotSettingResponse> chatbotSetting(
             @Valid ChatbotSettingRequest request,
-            CustomUserDetails userDetails
-    );
-
-    @Operation(summary = "자립목표영역 선정", description = "자립목표영역설정을 요청합니다.")
-    @ApiResponse(description = "목표설정 성공", responseCode = "200")
-    ResponseEntity<MissionAreaSettingResponse> missionAreaSetting(
-            @Valid MissionAreaSettingRequest request,
             CustomUserDetails userDetails
     );
