+++ src/main/java/com/dash/leap/domain/user/exception/UserExceptionHandler.java
@@ -19,12 +19,4 @@ public class UserExceptionHandler {
         log.info("InvalidChatbotTypeExceptionResponse: {}", exceptionResponse);
         return ResponseEntity.badRequest().body(exceptionResponse);
     }
-
-    @ExceptionHandler(InvalidMissionAreaChangeException.class)
-    @ApiResponse(responseCode = "400")
-    public ResponseEntity<ExceptionResponse> handleInvalidMissionAreaChangeException(InvalidMissionAreaChangeException e) {
-        ExceptionResponse exceptionResponse = new ExceptionResponse(HttpStatus.BAD_REQUEST.toString(), e.getMessage());
-        log.info("InvalidMissionAreaChangeExceptionResponse: {}", exceptionResponse);
-        return ResponseEntity.badRequest().body(exceptionResponse);
-    }
 }
